fix(agents): address skill menu review feedback
