+++ app/src/components/agent/SkillPromptInputBoundary.tsx
@@ -1,11 +1,13 @@
 import { Suspense, useEffect } from "react";
 import { ErrorBoundary } from "react-error-boundary";
 
-import type { AvailableAgentSkill } from "@phoenix/agent/skills/availability";
 import { PromptInputTextarea } from "@phoenix/components/ai/prompt-input";
 
 import { SkillPromptInput } from "./SkillPromptInput";
-import { useAvailableAgentSkills } from "./useAvailableAgentSkills";
+import {
+  useAvailableAgentSkills,
+  type AvailableAgentSkill,
+} from "./useAvailableAgentSkills";
 
 type SkillPromptInputBoundaryProps = {
   placeholder?: string;
