+++ app/src/components/agent/usePromptSkillCommand.ts
@@ -1,7 +1,8 @@
 import { useRef, useState } from "react";
 
-import type { AvailableAgentSkill } from "@phoenix/agent/skills/availability";
 import { findSkillTokens } from "@phoenix/agent/skills/requestedSkills";
+
+import type { AvailableAgentSkill } from "./useAvailableAgentSkills";
 
 /**
  * Describes the active slash-command query the user is typing, derived from the
